[syft] move service/context.py to types
