+++ packages/syft/src/syft/external/oblv/oblv_service.py
@@ -28,18 +28,18 @@ from ...serde.deserialize import _deserialize as deserialize
 from ...serde.serializable import serializable
 from ...service.code.user_code import UserCode
 from ...service.code.user_code import UserCodeStatus
-from ...service.context import AuthedServiceContext
-from ...service.context import ChangeContext
 from ...service.response import SyftError
 from ...service.service import AbstractService
 from ...service.service import service_method
-from ...service.user.user_roles import GUEST_ROLE_LEVEL
 from ...store.document_store import DocumentStore
+from ...types.context import AuthedServiceContext
+from ...types.context import ChangeContext
 from ...types.credentials import SyftSigningKey
 from ...types.credentials import SyftVerifyKey
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
 from ...types.syft_object import SyftObject
 from ...types.uid import UID
+from ...types.user_roles import GUEST_ROLE_LEVEL
 from ...util.util import find_available_port
 from .constants import DOMAIN_CONNECTION_PORT
 from .constants import LOCAL_MODE
