+++ packages/syft/src/syft/service/__init__.py
@@ -1,8 +1,4 @@
 # relative
-from .context import AuthedServiceContext
-from .context import ChangeContext
-from .context import NodeServiceContext
-from .context import UnauthedServiceContext
 from .response import SyftAttributeError
 from .response import SyftError
 from .response import SyftException
