+++ packages/syft/src/syft/service/message/message_service.py
@@ -5,16 +5,16 @@ from typing import Union
 # relative
 from ...serde.serializable import serializable
 from ...store.document_store import DocumentStore
+from ...types.context import AuthedServiceContext
 from ...types.uid import UID
+from ...types.user_roles import DATA_SCIENTIST_ROLE_LEVEL
 from ...util.telemetry import instrument
-from ..context import AuthedServiceContext
 from ..response import SyftError
 from ..response import SyftSuccess
 from ..service import AbstractService
 from ..service import SERVICE_TO_TYPES
 from ..service import TYPE_TO_SERVICE
 from ..service import service_method
-from ..user.user_roles import DATA_SCIENTIST_ROLE_LEVEL
 from .message_stash import MessageStash
 from .messages import CreateMessage
 from .messages import LinkedObject
