+++ packages/syft/src/syft/service/network/node_peer.py
@@ -8,12 +8,12 @@ from typing_extensions import Self
 # relative
 from ...client.client import SyftClient
 from ...serde.serializable import serializable
+from ...types.context import NodeServiceContext
 from ...types.credentials import SyftSigningKey
 from ...types.credentials import SyftVerifyKey
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
 from ...types.syft_object import SyftObject
 from ...types.uid import UID
-from ..context import NodeServiceContext
 from ..metadata.node_metadata import NodeMetadata
 from .routes import NodeRoute
 from .routes import connection_to_route
