+++ packages/syft/src/syft/service/policy/policy_service.py
@@ -5,8 +5,8 @@ from typing import Union
 # relative
 from ...serde.serializable import serializable
 from ...store.document_store import DocumentStore
+from ...types.context import AuthedServiceContext
 from ...types.uid import UID
-from ..context import AuthedServiceContext
 from ..response import SyftError
 from ..response import SyftSuccess
 from ..service import AbstractService
