+++ packages/syft/src/syft/service/project/project_service.py
@@ -6,9 +6,10 @@ from typing import Union
 from ...serde.serializable import serializable
 from ...store.document_store import DocumentStore
 from ...store.linked_obj import LinkedObject
+from ...types.context import AuthedServiceContext
 from ...types.uid import UID
+from ...types.user_roles import GUEST_ROLE_LEVEL
 from ...util.telemetry import instrument
-from ..context import AuthedServiceContext
 from ..message.message_service import CreateMessage
 from ..message.message_service import MessageService
 from ..response import SyftError
@@ -18,7 +19,6 @@ from ..service import AbstractService
 from ..service import SERVICE_TO_TYPES
 from ..service import TYPE_TO_SERVICE
 from ..service import service_method
-from ..user.user_roles import GUEST_ROLE_LEVEL
 from ..user.user_service import UserService
 from .project import Project
 from .project import ProjectEvent
