+++ packages/syft/src/syft/service/user/__init__.py
@@ -6,11 +6,5 @@ from .user import UserPrivateKey
 from .user import UserSearch
 from .user import UserUpdate
 from .user import UserView
-from .user_roles import DATA_OWNER_ROLE_LEVEL
-from .user_roles import DATA_SCIENTIST_ROLE_LEVEL
-from .user_roles import GUEST_ROLE_LEVEL
-from .user_roles import ROLE_TO_CAPABILITIES
-from .user_roles import ServiceRole
-from .user_roles import ServiceRoleCapability
 from .user_service import UserService
 from .user_stash import UserStash
