+++ packages/syft/tests/syft/api_test.py
@@ -10,7 +10,7 @@ import pytest
 import syft as sy
 from syft.service.response import SyftAttributeError
 from syft.service.user.user import UserUpdate
-from syft.service.user.user_roles import ServiceRole
+from syft.types.user_roles import ServiceRole
 
 
 def test_api_cache_invalidation(worker):
