+++ packages/syft/tests/syft/messages/message_service_test.py
@@ -4,7 +4,6 @@ from result import Err
 from result import Ok
 
 # syft absolute
-from syft.service.context import AuthedServiceContext
 from syft.service.message.message_service import MessageService
 from syft.service.message.message_stash import MessageStash
 from syft.service.message.messages import CreateMessage
@@ -14,6 +13,7 @@ from syft.service.response import SyftError
 from syft.service.response import SyftSuccess
 from syft.store.document_store import DocumentStore
 from syft.store.linked_obj import LinkedObject
+from syft.types.context import AuthedServiceContext
 from syft.types.credentials import SyftSigningKey
 from syft.types.credentials import SyftVerifyKey
 from syft.types.datetime import DateTime
