+++ packages/syft/tests/syft/users/fixtures.py
@@ -3,20 +3,20 @@ import pytest
 
 # syft absolute
 from syft.node.worker import Worker
-from syft.service.context import AuthedServiceContext
-from syft.service.context import NodeServiceContext
-from syft.service.context import UnauthedServiceContext
 from syft.service.user.user import User
 from syft.service.user.user import UserCreate
 from syft.service.user.user import UserPrivateKey
 from syft.service.user.user import UserSearch
 from syft.service.user.user import UserUpdate
 from syft.service.user.user import UserView
-from syft.service.user.user_roles import ServiceRole
 from syft.service.user.user_service import UserService
 from syft.service.user.user_stash import UserStash
 from syft.store.document_store import DocumentStore
+from syft.types.context import AuthedServiceContext
+from syft.types.context import NodeServiceContext
+from syft.types.context import UnauthedServiceContext
 from syft.types.credentials import UserLoginCredentials
+from syft.types.user_roles import ServiceRole
 
 
 @pytest.fixture(autouse=True)
