+++ packages/syft/tests/syft/worker_test.py
@@ -15,7 +15,6 @@ from syft.client.api import SyftAPICall
 from syft.node.worker import Worker
 from syft.service.action.action_object import ActionObject
 from syft.service.action.action_store import DictActionStore
-from syft.service.context import AuthedServiceContext
 from syft.service.queue.queue_stash import QueueItem
 from syft.service.response import SyftAttributeError
 from syft.service.response import SyftError
@@ -23,6 +22,7 @@ from syft.service.user.user import User
 from syft.service.user.user import UserCreate
 from syft.service.user.user import UserView
 from syft.service.user.user_service import UserService
+from syft.types.context import AuthedServiceContext
 from syft.types.credentials import SIGNING_KEY_FOR
 from syft.types.credentials import SyftSigningKey
 from syft.types.credentials import SyftVerifyKey
